Move FA kit-specific stuff to the FA icon library
