+++ src/webawesome.ts
@@ -1,6 +1,7 @@
+export { getKitCode, setKitCode } from './components/icon/library.default.js';
 export { registerIconLibrary, unregisterIconLibrary } from './components/icon/registry.js';
 export { discover, preventTurboFouce, startLoader, stopLoader } from './utilities/autoloader.js';
-export { getBasePath, getKitCode, setBasePath, setKitCode } from './utilities/base-path.js';
+export { getBasePath, setBasePath } from './utilities/base-path.js';
 export { allDefined } from './utilities/defined.js';
 export { registerTranslation } from './utilities/localize.js';
